+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html) since [v0.1.0].
 
 ## [Unreleased]
+### Fixed
+- The calendar button of `DatePickerField` is also disabled when the field is.
 
 ## [v0.7.0] - 2020-10-20
 ### Added